Prove the first owner can really sign in
"The owner they made can sign in" was checking the address the browser landed
on, and a refused password lands on the same address as an accepted one. The
story passed with a completely wrong password. It now checks the site stopped
asking for one.

A refusal that made the owner anyway and then stopped short would leave the
site impossible to set up — the name taken, nobody able to sign in — and both
of the old checks passed in that state. The story now sets the site up properly
afterwards, which is the thing that would break.

Opening the setup page again is read as where the browser ended up, not only as
a missing button, and the rule about how short a value may be moved to its own
test file.
