+++ test/specs/support/form-controls.test.ts
@@ -287,43 +287,6 @@ describe("the days a page has ticked", () => {
     });
   });
 
-  describe("a box that asks for a minimum length", () => {
-    const password = (extra = "") =>
-      `<input name="pass" type="password" minlength="8"${extra}>`;
-
-    test("accepts a value long enough for the box", () => {
-      expect(whyValueCannotBeSent(password(), "pass", "longenough")).toBeNull();
-    });
-
-    test("refuses a value shorter than the box asks for", () => {
-      expect(whyValueCannotBeSent(password(), "pass", "short")).toBe(
-        "the pass box takes nothing shorter than 8 characters",
-      );
-    });
-
-    test("accepts a value of exactly the minimum", () => {
-      expect(whyValueCannotBeSent(password(), "pass", "12345678")).toBeNull();
-    });
-
-    // An empty required box is already answered by the required rule, which
-    // says something more useful than "too short".
-    test("leaves an empty required box to the required rule", () => {
-      expect(whyValueCannotBeSent(password(" required"), "pass", "")).toBe(
-        "the pass box must be filled in",
-      );
-    });
-
-    test("ignores length on a box that asks for no minimum", () => {
-      expect(
-        whyValueCannotBeSent(
-          '<input name="pass" type="password">',
-          "pass",
-          "x",
-        ),
-      ).toBeNull();
-    });
-  });
-
   describe("filling a form in and sending it", () => {
     /** A browser stand-in: it holds the page it was served and remembers what
      * was sent, so the helper's own two jobs can be seen separately. */
